+++ src/main/java/com/marcospassos/phpserializer/SerializerBuilder.java
@@ -3,7 +3,7 @@ package com.marcospassos.phpserializer;
 import com.marcospassos.phpserializer.adapter.ArrayAdapter;
 import com.marcospassos.phpserializer.adapter.BooleanAdapter;
 import com.marcospassos.phpserializer.adapter.IntegerAdapter;
-import com.marcospassos.phpserializer.adapter.ListAdapter;
+import com.marcospassos.phpserializer.adapter.CollectionAdapter;
 import com.marcospassos.phpserializer.adapter.MapAdapter;
 import com.marcospassos.phpserializer.adapter.ObjectAdapter;
 import com.marcospassos.phpserializer.adapter.ReferableObjectAdapter;
@@ -15,6 +15,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 
 /**
  * Provides a friendly API for creating instances of {@link Serializer}.
@@ -165,7 +166,8 @@ public class SerializerBuilder
         registerAdapter(char[].class, arrayAdapter);
         registerAdapter(short[].class, arrayAdapter);
         registerAdapter(Map.class, new MapAdapter());
-        registerAdapter(List.class, new ListAdapter());
+        registerAdapter(List.class, new CollectionAdapter());
+        registerAdapter(Set.class, new CollectionAdapter());
         registerAdapter(Boolean.class, new BooleanAdapter());
         registerAdapter(Integer.class, new IntegerAdapter());
         registerAdapter(String.class, new StringAdapter());