+++ CHANGELOG.md
@@ -4,11 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Versions
-
-### Unreleased
+## Unreleased
 There are currently no unreleased changes.
 
-### [0.1.0](https://github.com/marcospassos/java-php-serializer/releases/tag/0.1.0) (2017-06-29)
+## [0.2.0](https://github.com/marcospassos/java-php-serializer/releases/tag/0.2.0) (2017-07-06)
+
+### Changed
+
+- Rename `ListMap` to `CollectionAdapter` since it supports any collection.
+
+### Added
+
+- Adapter for `Set` in the list of builtin adapters registered by 
+`SerializerBuilder::registerBuiltinAdapters()`. 
+
+## [0.1.0](https://github.com/marcospassos/java-php-serializer/releases/tag/0.1.0) (2017-06-29)
 
 - First beta release